Добавляет хок для плеера (без функционала)
