+++ src/hocs/with-full-video-player/with-full-video-player.js
@@ -1,120 +1,32 @@
 import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 
-const MAX_PROGRESS = 100;
-const SEC = 1;
-
 const withFullVideoPlayer = (Component) => {
   class WithFullVideoPlayer extends PureComponent {
     constructor(props) {
       super(props);
 
       this.state = {
-        isPlaying: false,
-        progress: 0,
-        togglerPosition: 0,
-        seconds: 0,
-        time: 0,
-        interval: null
+        isPlaying: false
       };
 
-      this._videoRef = React.createRef();
-      this.onControlButtonClick = this._onControlButtonClick.bind(this);
-      this.onFullModeButtonClick = this._onFullModeButtonClick.bind(this);
+      this.onPlayButtonClick = this._onPlayButtonClick.bind(this);
     }
 
     render() {
       return <Component
         {...this.props}
         isPlaying={this.state.isPlaying}
-        progress={this.state.progress}
-        togglerPosition={this.state.togglerPosition}
-        seconds={this.state.seconds}
-        time={this.state.time}
-        videoRef={this._videoRef}
-        onControlButtonClick={this.onControlButtonClick}
-        onFullModeButtonClick={this.onFullModeButtonClick}
+        onPlayButtonClick={this.onPlayButtonClick}
       />;
     }
 
-    componentDidMount() {
-      const video = this._videoRef.current;
-
-      if (this.props.isPlaying) {
-        video.play();
-      }
-
-      if (video) {
-        video.addEventListener(`loadedmetadata`, () => {
-          this.setState({
-            time: video.duration
-          });
-        });
-      }
+    _onPlayButtonClick() {
+      this.props.onPlayButtonClick(!this.props.fullVideoShown);
 
       this.setState({
-        isPlaying: this.props.isPlaying
+        isPlaying: !this.state.isPlaying,
       });
-
-      this.runPlayer();
-    }
-
-    componentDidUpdate() {
-      const video = this._videoRef.current;
-
-      if (this.state.progress === MAX_PROGRESS) {
-        video.load();
-        this.stopPlayer();
-      }
-    }
-
-    _onControlButtonClick() {
-      const video = this._videoRef.current;
-
-      this.setState({
-        isPlaying: !this.state.isPlaying
-      });
-
-      if (!this.state.isPlaying) {
-        video.play();
-        this.runPlayer();
-      } else {
-        video.pause();
-        this.stopPlayer();
-      }
-    }
-
-    _onFullModeButtonClick() {
-      const video = this._videoRef.current;
-
-      video.requestFullscreen();
-    }
-
-    runPlayer() {
-      const intervalId = setInterval(() =>
-        this.setState({
-          seconds: this.state.seconds + SEC,
-          progress: this.state.progress + MAX_PROGRESS / this.state.time,
-          togglerPosition: this.state.togglerPosition + MAX_PROGRESS / this.state.time
-        }), 1000);
-
-      this.setState({
-        interval: intervalId
-      });
-    }
-
-    stopPlayer() {
-      clearInterval(this.state.interval);
-
-      if (this.state.progress === MAX_PROGRESS && this.state.isPlaying) {
-        this.setState({
-          isPlaying: false,
-          progress: 0,
-          togglerPosition: 0,
-          seconds: 0,
-          time: null
-        });
-      }
     }
   }
 
@@ -128,16 +40,8 @@ const withFullVideoPlayer = (Component) => {
       /* eslint-enable */
     }),
     isPlaying: PropTypes.bool,
-    format: PropTypes.string,
-    progress: PropTypes.number,
-    togglerPosition: PropTypes.number,
-    seconds: PropTypes.number,
-    time: PropTypes.number,
-    interval: PropTypes.number,
-    videoRef: PropTypes.object,
-    onPlayButtonClick: PropTypes.func,
-    onControlButtonClick: PropTypes.func,
-    onFullModeButtonClick: PropTypes.func
+    fullVideoShown: PropTypes.bool,
+    onPlayButtonClick: PropTypes.func
   };
 
   return WithFullVideoPlayer;
